Add node_bitmap_preempt field to job_record_t
This added bitstring will be used for cases where we are trying
to schedule jobs via preemption. In those cases, these jobs will
not have a job_record_t->job_resrcs->node_bitmap, which is needed
if we want to accurately determine if a job will not break accounting
policy limits. We will use this new field exclusively for that.

Ticket: 20890