Use job_record_t->node_bitmap_preempt for preemptor jobs
In cases where we need to evaluate if a preemptor job will
break accounting policy limits, we need to know which nodes would be
assigned so we can determine if there will be node overlap with
the currently qos (or similar) nodelist.

Normally, this is done using node bitmaps, but these specific types
of jobs get their job_record_t->job_resrcs->node_bitmap cleared
before they get to the function that calculates node overlap.

This commit makes sure that we save a copy of that information before
it gets cleared, and later used for the node overlap calculation.

Ticket: 20890
Cherry-picked: 283fbde11d