Prevent crash - Set reservation partition as well as part_ptr
When slurmctld restarts, resv_ptr->part_ptr is set based on
resv_ptr->partition in _validate_one_reservation(). If the
resv_ptr->partition is not set, then resv_ptr->part_ptr is not set.
This can lead to a crash in certain situations where the reservation is
expected to have a non-NULL partition and part_ptr.

Ticket: 21836
Cherry-picked: 482f98090f