Always aggregate the step->jobacct to external step stats
This way we keep previous data and when there are no pids in the extern
step we can still get some past statistics with sstat.

If there have not ever been any pid in the extern, then no stats will
be shown for the extern step in a sstat request.

Issue: 50129
Ticket: 12407