Prevent lower priority job running after cancelling an ASAP reboot
When cancelling the ASAP reboot, the node was still out of the
avail_node_bitmap until the next registration. If the backfill
scheduler happened before the registration, the job would be scheduled.
So when cancelling an ASAP reboot, put the node back in the
avail_node_bitmap to close the race.

Changelog: Prevent lower priority job running after cancelling an ASAP
 reboot.
Ticket: 18087
Cherry-picked: fb2508018d