Revert commits 3669b954c35 and adbb98eb08d which introduce a regression
When requesting a job with a specific number of nodes, tasks,
cpus-per-task > 1 and a --mem-per-cpu value (or having DefMemPerCpu),
if -N != 1 and tasks/nodes > 1, controller won't be able to find a valid
node configuration. E.g:

$ srun -N 2 -n 4 -c 2 --mem-per-cpu=200 hostname
srun: error: Unable to allocate resources: Requested node configura..

This regression was originally introduced to address ticket 21255. The
following commit contains the alternate approach to fix that issue.

Ticket: 21255
Ticket: 22026
Changelog: Fix certain job requests that were incorrectly denied with
 node configuration unavailable error.