Prevent jobs with --gpus to be rejected in heterogenous partitions
Jobs submitted with --gpus were rejected in partitions where nodes had
different number of resources (GPUs and CPUS) configured. Asking for an
amount of GPUs that only a subset of nodes could supply was failing when
combined with some values of --cpus-per-task.

This is an alternate approach to 3669b954c35 and adbb98eb08d , which were
reverted in the previous commit, as they were introducing a regression.

Ticket: 21255
Ticket: 22026