Add flag to preserve FUTURE node state on restart
Add ReconfigFlags=KeepNodeStateFuture. If a node is FUTURE in slurm.conf,
restore saved node state if there is any. This used to be preserved on
reconfig, but it needs to be restored from state now.

This will be made the default in 25.05.

Changelog: Add ReconfigFlags=KeepNodeStateFuture to restore saved FUTURE
  node state on restart and reconfig instead of reverting to FUTURE
  state. This will be made the default in 25.05.
Ticket: 22027