Requeue resume failed jobs
If a cloud node fails to be resumed, the associated jobs are requeued if
they have the requeue flag. This commits adds a SchedulerParameter
called "requeue_on_resume_failure" that forces the jobs to be requeued
in this situation even if they do not have the requeue flag.

Issue: 50309
Ticket: 19306
Changelog: Add new SchedulerParameter option to force job requeue in the
 situation that a cloud node failed to be resumed.