data_parser/v0.0.40 - Fix memory leak when parsing list_t *
Before in _parse_list() if the dst list was NULL it would create a new list
with parser->free as the list destruction function. However parser->free is
always NULL since the addpl() macro does not allow for a free function to
be specified.

This changes _parse_list() to use the free function of the parser it is a
list of.

Ticket: 22205
Changelog: slurmrestd - Fix possible memory leak when parsing arrays with
 data_parser/v0.0.40.
Cherry-picked: 3051b9af72