Add ability to get ungrouped assoc usage
For some reports it is desirable to see the usage broken out by
individual assocs for a user including those that are part of a
partition, while for others we want to concatenate usage onto higher
level accounts and users. We adjusted the sreport usage query to allow
both report types to group usage differently.

Ticket: 20494