Fix padding parsing when expanding stdio fields
When expanding stdio patterns we are not clearing the padding from
a previous state transition, thus if printing two numeric fields
the second one will inherit the padding of the first one if there is no
new padding setting.

We also need to fix a strange behavior that is in slurm which removes all
the digits except for one when the padding is greater than 10 and the next
character is not a wildcard.

Changelog: Fix padding parsing when printing stdio fields for jobs.
Issue: 50279