Prevent deletions of a WCKey if a job is running with it
Originally, WCKeys could be deleted even if a job was running on the WCKey.
WCKeys will now have the same behavior as other delete-able entities, where if
there are jobs running on the entities, the entity will not be deleted.

Changelog: Prevent deletions of a WCKey if a job is running with it.
Ticket: 20854