backfill - minor performance enahcements
The licenses_unavail flag can't be unset, so further testing is useless.
No matter what, we will SKIP_SCHED_OR_TRY_LATER(). The benefit of the
current behavior is that we could move the later_start future in one
TRY_LATER iteration, which could be better depending on the situation.

Issue: 50278