Remove unnecessary condition for ntasks_per_node in job_limits_check
This removes an if-else block introduced in commit d413c8b7. The else
block was always a no-op since detail_ptr->ntasks_per_node is never
expected to be NO_VAL16. If detail_ptr->ntasks_per_node is
non-zero it always sets detail_ptr->num_tasks. Therefore, we can remove
the if-else and assign the value directly.

Ticket: 22163