Change from first_step_name to first_step_id in job_stp_t
This was considering the first step id of the job as a char *, with a
confusing name of first_stepname. Change the type to a uint32_t as it is
just the real step_id, and use also the numeric step for SBATCH step.

This allows the pattern expansion to work with others than the batch step
and does not hardcodes "batch" anymore.

Note that the job will always be the one from the batch.

With this change we also allow to use padding for numeric steps, which will
be needed when printing expanded fname patterns from these steps.

This is in preparation for next commits.

Issue: 50279