Implement logic for license OR'ing in job requests
Allocate any one of the licenses that were requested. Track the allocated
licenses in a separate field and update job_ptr->license_list to the
allocated license(s) when the job is allocated. This makes it easy to get
and return the correct licenses from cluster_license_list when the job is
allocated, completed, or state saved/restored. This also automatically gives
us the correct values in ReqTres and AllocTres if licenses are tracked in
AccountingStorageTres.

There is an important limitation in this implementation:

We always attempt to allocate the first license that is available in the
order that they were requested. This has important implications for
reservations that have licenses. For example:

* License A: 10
* License B: 20
* A reservation has 5 of license A and 10 of license B
* A job requests the reservation and --licenses='A:5|B:10'. The job will be
  allocated 5 of license A.
* The same job request comes in. The job will be allocated 5 of license A,
  even though there are 10 of license B in the reservation that are unused.
  This results in unoptimal use of licenses: jobs outside of the reservation
  are unable to use license A, while there are unused license B in the
  reservation.

This limitation occurs because of how licenses are implemented in
reservations: We track how many licenses are reserved (if someone calls
scontrol show licenses, although that can easily be fixed). However, we
don't track how many licenses are used from within a reservation. We only
track how many licenses are used in total. Therefore, when we checking if
licenses are available to the job, we currently can't distinguish between
licenses inside and outside of a reservation.

Issue: 50203
Ticket: 20934