Prevent jobs using --cpus-per-gpu and --mem keep pending forever
When a job requests --mem and MaxMemPerCPU is in place, we potentially
adjust the CPU values in _valid_pn_min_mem. If the CPUs were requeted
with --cpus-per-gpu, the topology plugins will try to use the value
requested instead of the calculated one. The changes ensure that we end
up using the upscaled value.

Ticket: 22018