Set CPU values for jobs requesting tasks and --mem
When a job requests --mem and MaxMemPerCPU is in place, we potentially
adjust the CPU values in _valid_pn_min_mem. The commit updates the CPUs
per task for jobs requesting tasks and nodes, so the job does not
violate the MaxMemPerCPU in place. It aims to produce a job with the
least amount of CPUs possible.

Ticket: 22018