Enable license OR with bf_licenses
Enable license OR when backfill planning for licenses. The logic is similar
as testing/allocating licenses for jobs. When testing for availability with
OR, any one of the licenses is sufficient. When deducting licenses for
backfill planning, find the first license where we have a sufficient amount.

We do not need to modify bf_license_transfer because that is only used for
reservations, not for jobs.

This implementation to backfill planning has the same limitation as job
allocation in the previous commit: OR'd licenses are planned in the order in
which they are requested, and this could result in unoptimal planning for
licenses. See the previous commit for a more detailed explanation.

Issue: 50203
Ticket: 20934