Revert "slurmctld - Fix issue where failed hetjob could crash slurmctld"
This reverts commit 8ececcd99b1ff79fb6191a44cbddd42c64c3a555.

Previous list_for_each() calls needed to be replaced with iterators
when canceling a hetjob since nested ones were not safe to use
in the same thread. Revert these since commit b1eb96daca enables
list_for_each() to be nested within the same thread.

This is a continuation of 8ececcd99b.

Ticket: 22103