Docs - Clarify cgroup v1 vs v2 resource specialization behavior
In cgroup/v1 the slurmstepd is constrained by the same memory and cpuset
limits as the slurmd if MemSpecLimit and or CoreSpecCount/CpuSpecList are
configured. In cgroup/v2 the slurmstepd is always contrained by the job's
cgroup since it is considered part of the job and its consumption is
completely dependent on the topology of the job.

Ticket: 22466