slurmstepd - Add --wait-for-children validation
srun is already validating this so this path should not typically be
hit, but we still need validate on slurmstepd as well in case of a
modified/buggy client.

In the case of checking cgroup vserion though, we need to validate on
slurmstepd since srun does not know what cgroup version is running.

Issue: 50219