data_parser/v0.0.42 - Add FLAG_INLINE_ENUMS to not $ref enum arrays
In 24.05, enum arrays were always inlined in the generated OpenAPI spec.
However, commit 214c824f caused enum arrays to be referenced. This was an
issue for oapi-codegen because it has some issues currently where
referenced objects that contain enums are not generating in the Golang
client correctly. This patch adds the data_parser flag inline_enums to
v0.0.42, which can be used to inline enum arrays in the spec.

Issue: 50297
Changelog: data_parser/v0.0.42 - Added +inline_enums flag which modifies
 the output when generating OpenAPI specification. It causes enum arrays to
 not be defined in their own schema with references ($ref) to them. Instead
 they will be dumped inline.