task/cgroup - Check cgroup_g_constrain_apply rc
The return code for cgroup_g_constrain_apply was not checked, so a job that
failed to apply any device constrain continued on its merry way.

We do not want a job which required constrain over a device to run
unconstrained. This commit will cause the step to fail when starting up.

Changelog: Ensure that a device constrain has been successfully applied to
 the job/step cgroup, so jobs do not run unconstrained after a failure.
Ticket: 22488