Testsuite - Fix 112_40 using fixtures ensuring proper cleanup
Several test cases create users, accounts and wckeys. If they are not
properly removed, they may interfere with the next test cases.
For example the first slurmdb_vxxxx_post_users() in test_db_users will be
treated like an update instead of a create which leads to issue with
wckeys.

Ticket: 18939