cgroup/v2 - Implement cgroup_p_get_task_empty_event_path
In a later commit, an inotify instance will be created which will watch
for IN_CLOSE events on a file which should indicate that we should check
if the task cgroup is empty or not. cgroup/v2 implementation is straight
forward because cgroup.events is meant to be watched with inotify
according to the cgroup man page.

Issue: 50219