Change return conditions in uid_from_string() when given non-existent uid
Return ESLURM_USER_ID_UNKNOWN and a pointer to the numeric
uid from the given user name string if the uid does not exist on
the system.

This will be used by squeue and scancel to help identify old jobs in
queue where the user has been removed from the underlying system.

Ticket: 21105