tls/s2n - Don't try to load CA cert in slurmstepd
When tls_enabled() is true, slurmstepd does not have to communicate
across the network to other daemons, so it does not need the CA cert
loaded.

Right now, slurmstepd doesn't have slurm_conf.tls_params because slurmd
is not packing it when sending to slurmd. It could be easily added, but
it isn't technically necessary, and this commit prevents any errors
trying to do _load_ca_cert without slurm_conf.tls_params.

Issue: 50070