tls/s2n - Change file permission check from fatal() to error()
This will allow slurmrestd to try to load tls/s2n for creating https://
sockets, but ignore any errors (as long as tls/s2n is not configured in
slurm.conf).

When tls/s2n fails to load in every other case (which will only happen
if tls/s2n is configured in slurm.conf), this will still eventually
result in a fatal().

Issue: 50293