tls/s2n - Add ephemeral certificate support to tls_p_create_conn()
In some cases, we want to connect to a server that does not have a
certificate signed by a CA in our trust store (e.g. slurmstepd
connection to srun for step IO).

We do not want to add ephemeral certificates to our default trust store,
so in such cases, we create a new ephemeral s2n config that only has the
ephemeral certificate in its trust store.

Issue: 50070