Send client TLS cert to slurmctld in slurm_allocate_resources_blocking()
If slurmctld cannot provide an immediate allocation,
slurm_allocate_resources_blocking() will open a listening port that
slurmctld will eventually need to connect to.

To open a connection on this port in a TLS environment, slurmctld needs
to have the TLS cert associated with the listening port opened by
whoever called slurm_allocate_resources_blocking().

This certificate will be stored on the job record in the next commit,
and will be used for other client connections (such as x11 forwarding)
later.

Issue: 50070