Move check for registered clusters
The if/else block this check is being moved from is primarily for
checking for running/completed jobs. By moving this code, the previous
if/else block can keep that simple purpose, and the check for registered
clusters (which is only relevant for cluster_table) is easier to read.

Ticket: 20854