cgroup/v2 - Fix slurmd reconfig not working when removing CoreSpecLimits
If slurmd was started manually, with CoreSpecLimits set, and then the
limits were completely removed from slurm.conf and slurmd reconfigured
e.g. with a scontrol reconfig, slurmd would fail to start complaining
about ENOSPC.

This happens because the cpuset.cpus and cpuset.mems limits cannot be
reset by writting an empty string to the interface, if there is a process
in it (the slurmd itself).

The kernel seems to interpret the empty string as to remove all cpus
or mem nodes, even if the fact that cpuset.cpus/mems.effective shows all
the available cpus or mems when the interface is empty.

The solution is to explicitly specify all the cpus/mems available, so we
now read the cpuset cpus and mems effective from the parent cgroup and
apply them to the slurmd cgroup.

Regression caused in commit master commit 9be763a4 and in 24.11 2504bddb.

Changelog: Fix reconfigure failure of slurmd when it has been started
 manually and the CoreSpecLimits have been removed from slurm.conf.
Ticket: 20943
Cherry-picked: 986d0b5a9b