switch/hpe-slingshot - Allow the slurmctld to unpack switch stepinfo
Fixes a regression added in 5108bdedfa.

The slurmctld needs to unpack the step_ptr->switch_step from the save state
so that it can free a step's allocated VNI on REQUEST_STEP_COMPLETE when
the stepmgr is not enabled.

Ticket: 22602
Changelog: switch/hpe-slingshot - Make sure the slurmctld can free step
 VNIs after the controller restarts or reconfigures while the job is
 running.
Cherry-picked: 83708bead0