Change Unspecified address in _ip_reserved_to_str() per RFC6874
Change IPv6 unspecified address from "::" to "[::]" and loopback as
"localhost" to "[::1]" based on the options in RFC6874 and formatting in
RFC4291.

Avoid the ambiguous "::" vs ":::" vs ":::PORT" which can be parsed as IPv4
by some clients. The "[::]" is unambiguously an unspecified IPv6
address.

Avoid the ambiguous "localhost" which could be resolved as IPv4 or IPv6
depending on the host configuration. The "[::1]" is unambiguously the IPv6
loopback address.

Issue: 50377