switch/hpe_slingshot - Allow FM token to be renewed on 401 or 403 error
At init time, the Slurm Slingshot plugin uses a Fabric Manager (FM) login
endpoint to get a token used for authn/authz for subsequent REST calls to
the FM. When that token expires, the plugin is supposed to login again and
get a new token. However, when the token expires, the FM currently returns
a HTML error string instead of the expected JSON, so the response-to-JSON
routine fails without trying to re-acquire the token.

This patch moves the check for a HTTP_UNAUTHORIZED (401) and HTTP_FORBIDDEN
(403) status before the response_str is considered. This prevents it from
erroring out due to it not receiving the expected JSON output. It will now
always try to re-acquire the token on a HTTP 401/403 status.

Ticket: 22439
Changelog: switch/hpe_slingshot - Always try to renew the fabric manager
 token on a HTTP 401 or 403 error status regardless of the contents of the
 response body.