mpi/pmix - Inject PMIX_APPLDR to PMIx KVS
Add support for injecting this new key into the KVS datastore.T
According to the PMIx 5.0 specification, this is an application realm
attribute, which in Slurm terminology means it is a heterogeneous step
attribute and provides the lowest global rank of the specified component.

For example with srun -n2 : -n3 ./myapp, the app leaders will be 0 and 2.

Issue: 50072