Add and set new het_job_step_task_cnts field to the step records
In preparation for next commits, we need to store the relation between
the heterogeneous step and its number of tasks. At the moment we have
other relations like the tasks in each node of the het job, but nothing
that can provide us this information. This will be used to return the
first rank of each component.

Issue: 50072