Fix memory leak of switch_jobinfo in slurmctld, slurmd, and slurmstepd
The job's switch_jobinfo structure was only being freed by the slurmctld in
deallocate_nodes(). However, this struct is unpacked by slurmds and
slurmstepds. Additionly, the structure was not freed by the slurmctld if it
shutdown while the job was running. In these cases,
switch_g_job_complete(), which normally frees the structure, should not be
called since that deallocated the VNIs. So switch_g_free_jobinfo() was
added and called in job_record_delete() instead to just free the structure.

Ticket: 22572
Changelog: Fix slurmctld, slurmd, and slurmstepd memory leak resulting
 from the use of switch/hpe_slingshot with the stepmgr enabled or
 switch/nvidia_imex. The slurmctld and slurmstepd memory leaks only occur
 while the daemons were shutting down.