Setup logging earlier in slurmd initialization
Some plugins initialize before _read_config so they do not log to the
slurmd log file, as the setup of logging facilities is embedded into
this function.

In the next commit we are going to move the initialization of
cgroup_g_init before _read_config and as it will set up the slurmd cgroup
and the scope, having logs is a must. This will benefit other plugins that
were initializing before _read_config.

Issue: 50237