cgroup/v2 - Implement constrain_g_get for memory.max
We return the memory.max of the requested cgroup, or if set to 'max' we
look for the limit on each ancestor in the cgroup tree. Unlike cgroup/v1
the max is the limit imposed on this cgroup but does not tell us how much
memory we can use.

We will stop when we arrive to a cgroup which does not have the memory
interface, or when we succeed finding a memory.max with numeric content.

Preparation for next commits.

Issue: 50237