Add slurm_unpack_msg_and_forward()
This function is the same as slurm_receive_msg_and_forward(),
except it takes a 'buffer' with the message already received on it.

This is intended to be used by conmgr which will have a message ready in
a buffer, so it can unpack it and forward it easily.

Issue: 50342