certmgr - sign_csr with name instead of node record
sackd will need a signed certificate, but it does not exist in the node
record table. This commit modifies certmgr_g_sign_csr() to accept a name
which can either be a hostname or nodename.

Note that since there's no node record for sackd, and thus no way to set
the internal token, the validation script will be required to validate
sackd's node token.

Issue: 50039