tls/s2n - Remove connection lock
This was a good idea, but it introduces some bad race conditions, and
shouldn't be necessary.

If one thread is calling tls_g_destroy() and another thread calls
tls_g_send(), the tls_g_send() thread may end up waiting on the conn->lock
while tls_g_destroy() holds it. As soon as tls_g_destroy() releases it
however, it then destroys the lock and frees conn. There is no guarantee
that tls_g_send() will finish its operations on conn before this happens,
and so tls_g_send() may try to access conn after its already been freed.
And even if tls_g_send() gets to conn before it's freed, it may try to
unlock conn->lock after its been destroyed, which would also introduce
undefined behavior.

The conn->lock is trying to make certain operations on the s2n_conn
atomic when in reality the calling code should already be doing this
since the same atomic restrictions apply to a standard non-TLS
connections on raw fd's. That is, if there are any concurrency issues
for a connection, TLS or not, the code that surrounds it should already
be handling concurrency. Any internal concurrency handling would be
redundant.

Issue: 50039