tls/s2n - Split _add_ca_cert_to_client() into two functions
In an upcoming commit, a tls plugin function will be added to load CA
cert after init, which will take in a path to a CA cert. The function
split here will allow that commit to use the same logic to load the CA
cert from a file that's not configured in slurm.conf.

Issue: 50039