cgroup/v2 - Do not try to empty pids of a non-root cgroup in a container
If we have been started manually in a container, so we are in the root
cgroup, we consider we are in a fake-root cgroup environment. In that case
we will create a hierarchy of directories and move all pids in subdirs to
not violate the kernel/cgroup rules and avoid having processes in non-leaf
directories.

After a reconfigure, the slurmd will be in one of these subdirectories but
it will still be in the fake-root environment, so it will try to empty pids
again.

We need an extra check to see if we are in the top-level directory on the
cgroup hierarchy before attempting to create the hierarchy.

Ticket: 22618
Changelog: When running slurmd in a container and a reconfigure is issued,
 do not re-generate the cgroup hierarchy inside the actual cgroup, which
 would end up having unnecessary cgroup sub-levels.