Don't do graceful TLS shutdown when server closes first
When the server-side closes first, the client-side may be inactive and
not respond with it's own graceful TLS shutdown, meaning the server-side
will block.

In this case, we don't want to do a graceful TLS shutdown. The server
side simply closes its connection, and when the client next checks the
connection, it can see if the underlying connection is closed meaning a
graceful TLS shutdown is no longer necessary.

Issue: 50039