slurmctld - Add TLS support to _service_connection()
conmgr will detect incoming TLS connections on_fingerprint_tls, and
create a server TLS connection to negotiate with the client TLS
connection sending a request. If a Slurm message is successfully parsed,
we extract the fd from conmgr, as well as the TLS connection used to
receive the message.

We want to use that same fd and TLS connection to send responses back to
the incoming message, so it is set here.

In addition, the TLS connection needs to be cleaned up just before we
close the underlying connection.

Issue: 50039