stepd_proxy - Change rc_msg to be dynamically allocated
Later in _on_data_local_socket(), we call slurm_free_msg_data() on
resp.data. This is because resp.data could have been dynamically
allocated from _slurmd_send_recv_msg().

It is also possible, however, that resp.data is pointing to rc_msg,
which is a static struct in _on_data_local_socket(). We do not want
to try to free a static struct.

This commit changes rc_msg to be dynamically allocated so that when we
get to slurm_free_msg_data(), we always know that resp.data is either
NULL, or pointing to data that can be freed. Although we don't
technically need to dynamically allocate memory for rc_msg, it results
in a simpler solution.

Issue: 50039