Change federated multi-cluster submission to test all given clusters
Previously, a federated multi-cluster (e.g. -M/--clusters) submission
would only test one cluster in a federation. When this design was first
implemented the origin cluster requested will_runs to each cluster and
returned the fastest starting cluster back to the client.

This changes it so that each given cluster is tested, regardless of
being in a federation or not, allowing a multi-cluster submission to
pick the cluster in the federation that could run the job the soonest.
The behavior is not the same whether in a federation or not.

Changelog: Allow federated multi-cluster submissions to test each
 given cluster rather than one cluster per-federation.
Issue: 50260