cgroup/v2 - Fix scope path generation after reconf of a slurmd with pid 1
When a reconfigure to a running non-daemonized slurmd with pid 1 is issued,
the new slurmd will generate an incorrect cgroup path for the stepd.

This commit sets the stepd scope cgroup path to the parent cgroup path of
the new slurmd, which in effect will be the scope as slurmd has been moved
to slurmstepd.scope/slurmd/ directory at the first init.

This typically happens on containers, with a manual slurmd startup with -D
flag (no daemonize). Note that a reconfigure of a daemonized slurmd which
has pid 1 is not supported, as slurmd will fork itself into a new pid and
pid 1 will terminate. If pid 1 terminates, the system (or the container)
goes down.

Ticket: 22618
Changelog: Fix stepd scope path generation after reconfiguring a
 non-daemonized and manually-started slurmd that is pid 1. This might
 happen only in containers and would cause incorrect cgroup hierarchy
 to be created.