mpi/pmix - Fix not identifying a regular job vs an het job
An heterogeneous allocation followed by a single srun hostname would cause
to enter the code where pmix sets up structs for an heterogeneous job with
multiple components. A NO_VAL (-1) in step->het_job_offset identifies a
regular step. An INFINITE (-2) identifies an unitialized het_job_offset.

Translated to uint32_t, NO_VAL < INFINITE so we need a < check and not
a !=.

Issue: 50072