sackd - Fix binding to SlurmdPort instead of an ephemeral port by default
When --port / SACKD_PORT is not specified, SlurmdPort should be the port
used by default to listen for reconfiguration updates.

Commit 235339c0d12 introduced a regression which caused sackd to listen
for such updates on an ephemeral port instead of SlurmdPort, resulting
in slurmctld sending REQUEST_RECONFIGURE_SACKD to SlurmdPort while sackd
was listening on the ephemeral port instead, resulting in the loss of
such RPC requests.

Ticket: 22700