Testsuite - Adjust pack_cluster_rec unit test for previous commit
Test was faking the fd, but the pack now looks at tls_conn. Set tls_conn
to a non-NULL address to get the same behavior as before.

Remove the assertions on the fd in anticipation of these being removed
from the persistent connection structure in the future. The non-NULL
check on the structure is adequate to prove the underlying signal as
to whether the cluster is connected or not.