xcpuinfo - Fix _load_gres with p-cores and e-cores
Logic to remove e-cores from the range of xcpuinfo_mac_to_abs() should only
apply for generating the CpuSpecList, but not for _load_gres() or others
where the entire cpu list is needed.

Just move the code that must be specific to xcpuinfo_get_cpuspec() where
it corresponds.

This fixes and completes 4789d3ed96.

Issue: 50237