Fix coverity issue in node_conf
We were casting to int while all the variables are uint16_t. Note that
the loop might substract -1 if tasks per core is 0 and make the result
negative, which should never happen as this code should never run for
tpc < 2.

This popped up after the development of issue 50237 while moving code
around to make it common in 43f1689235, so it is not new. The previous
fix in 88ac8ab7a33 seems was not sufficient.

CID: 521389
Issue: 50237