cgroup - Make plugin thread-safe
It is possible that several threads come in a slurmstepd trying to add
or remove tasks from the external step. This can trigger a race which
will lead to a segfault or to different issues if a thread trying to
add an external pid, is moving it into a cgroup that is being destroyed.
This happens because proctrack works at task_special directory, but
other plugins work at task_x level, and when they initialize they
destroy the task_special cgroup structure which another thread might
still be using.

This is independent of common_cgroup_lock() that works at filesystem
level by trying to lock a cgroup to avoid other steps to access them.

We need to make the plugin be thread-safe by adding mutexes in every
operation where pid movements or counts are involved.

Note that this will mainly affect pam_slurm_adopt and the add extern pid
rpc, because we do not create other threads anywhere that make use of
cgroups, so performance should, for the moment, not be impacted.

Changelog: Remove possible race conditions when pids are orphaned in a
 pam_slurm_adopt ssh session and another incoming RPC is trying to add a
 ssh session into the extern cgroup. In general the plugin was not thread
 safe and the only way threads can be created and conflict is with the
 handle_add_extern_pid API function, that pam_slurm_adopt is using.
 This caused segfaults in extern slurmstepd in some specific cases.
Issue: 19333
Issue: 19451