Revert "Cleanup task_special when it is not used anymore in cgroup/v2"
This reverts commit 3e2bded95df4c2817024db636f3d9bfcce79c8f3.

The task_special directory was removed when another plugin moved the pids
from it to another task_x specific cgroup. But that meant that everytime
we called proctrack and cgroup_g_step_add, the task_special directory
would be created again. This adds overhead to cgroup and an call from
pam_slurm_adopt or any rpc that adds an extern pid to the step will create
this directory again.

The statement from the original commit message says that it is possible
that a systemctl daemon restart/reload deletes the empty cgroups from the
tree, but since slurmstepd scope will always start with delegate=yes that
should not happen.

This is a commit to improve performance and diminish the impact on the
cgroup hierarchy.

Ticket: 19333
Ticket: 19451