Fix x11 cleanup race condition
The thread used to clean the x11 files was dropping privileges. If in the
meantime a cgroup or other tasks from step cleanup were taking place and
needed privileges, they would fail with EPERM. By adding locking with
auth_setuid_lock we guarantee that while doing these cleanups the
privileges will not be dropped.

Other parts of the code like the removal of the domain socket were also
affected by this, so the locking has been extended also to these places.

Changelog: Fix race condition on x11 shutdown which caused other privileged
 cleanup operations to fail and leave stray cgroup or spool directories and
 errors in the logs.
Ticket: 22736