switch/hpe_slingshot - Prevent slurmctld crash to due NULL vnis bitmap
Before SwitchParameters=vnis=min-max was documented as being required for
switch/hpe_slingshot. However, the requirement was not enforced. If
SwitchParameters was specified but vnis was not, the vnis_table would be
left NULL. This would lead to segfaulting the slurmctld.

This makes the vnis parameter optional and has it default to 1024-65535.

See 2ccd4fe7fd4.

Changelog: switch/hpe_slingshot - Fix defaulting VNI range to 1024-65535
 when no SwitchParameters specified.