auth/slurm - Fix slurmd runtime directory permission checks
RUNTIME_DIRECTORY was only meant to be interpreted by sackd. If it was set
(which will always be if daemon started via systemd), then slurmd would
execute the case where the env var is parsed, passing true (SlurmUser)
as argument to _prepare_run_dir(), but it should be passing false
(SlurmdUser) and ignore the env var.

Continuation of c40e571c72.

Issue: 50346
Ticket: 20682