Make typed GRES reservations work without having to also include the non-typed GRES
If a reservation is requested without the typeless gres, it does not know
that the typeless gres is requested in addition to the typed gres. This
function checks if the typeless gres has not been added, then adds it if
needed.

Changelog: Make typed GRES reservations work without having to also
 include the non-typed GRES
Ticket: 18861