Gres reservations will work without requiring shard or mps
_can_use_gres_exc_topo() uses the i loop variable as gres_bit, but the i
variable was not always equal to gres_bit. This could happen if the gres
definition in the gres.conf was on a single line and shard or mps was not
specified as a gres type in the slurm.conf. When the i variable was not
equal to gres_bit, it was possible for the reservation to take up all gres
of that type. _can_use_gres_exc_topo() now loops through all gres on the
topology and increments the avail_gres if the gres is available to the job.

Changelog: Gres reservations now work as expected when shard is not defined
 as a GresType in the slurm.conf.
Ticket: 18861